remove client directive from top layout, refactor client pages, added sessionmodal on client
