+++ components/about-components/BrandSlider.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Autoplay } from "swiper/modules";
