+++ components/about-components/TeamSlider.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Pagination, Navigation } from "swiper/modules";
