--- context/ShoppingCartContext.tsx
@@ -1,80 +1,0 @@
-import { useContext, createContext, ReactNode, useState } from "react";
-
-type ShoppingCartProviderProps = {
-  children: ReactNode;
-};
-
-type CartItem = {
-  id: number;
-  quantity: number;
-};
-
-type ShoppingCartContextType = {
-  getItemQuantity: (id: number) => number;
-  increaseCartQuantity: (id: number) => void;
-  decreaseCartQuantity: (id: number) => void;
-  removeFromCart: (id: number) => void;
-};
-
-const ShoppingCartContext = createContext({} as ShoppingCartContextType);
-
-export function useShoppingCart() {
-  return useContext(ShoppingCartContext);
-}
-
-export function ShoppingCartProvider({ children }: ShoppingCartProviderProps) {
-  const [cartItem, setCartItem] = useState<CartItem[]>([]);
-
-  function getItemQuantity(id: number) {
-    return cartItem.find((item) => item.id === id)?.quantity || 0;
-  }
-
-  function increaseCartQuantity(id: number) {
-    setCartItem((currentItems) => {
-      if (currentItems.find((item) => item.id === id) == null) {
-        return [...currentItems, { id, quantity: 1 }];
-      } else {
-        return currentItems.map((item) => {
-          if (item.id === id) {
-            return { ...item, quantity: item.quantity + 1 };
-          } else {
-            return item;
-          }
-        });
-      }
-    });
-  }
-  function decreaseCartQuantity(id: number) {
-    setCartItem((currentItems) => {
-      if (currentItems.find((item) => item.id === id)?.quantity === 1) {
-        return currentItems.filter((item) => item.id !== id);
-      } else {
-        return currentItems.map((item) => {
-          if (item.id === id) {
-            return { ...item, quantity: item.quantity - 1 };
-          } else {
-            return item;
-          }
-        });
-      }
-    });
-  }
-
-  function removeFromCart(id: number) {
-    setCartItem((currentItems) => {
-      return currentItems.filter((item) => item.id !== id);
-    });
-  }
-  return (
-    <ShoppingCartContext.Provider
-      value={{
-        getItemQuantity,
-        increaseCartQuantity,
-        decreaseCartQuantity,
-        removeFromCart,
-      }}
-    >
-      {children}
-    </ShoppingCartContext.Provider>
-  );
-}
