+++ data/review.data.ts
@@ -1,0 +1,65 @@
+export const reviews = [
+  {
+    client: "Emily Rodriguez",
+    title: "Heavenly Indulgence on a Plate!",
+    comment:
+      "Indulging in Goldis Cake is like taking a trip to dessert paradise. The cakes are moist, the toppings are scrumptious, and every bite is a taste sensation. Pure bliss!",
+    platform: "twitter",
+  },
+  {
+    client: "David Patel",
+    title: "Irresistibly Tempting Treats!",
+    comment:
+      "Goldis Cake crafts cakes that are impossible to resist. From the first glance to the last crumb, their creations are a feast for the eyes and the taste buds. Simply irresistible!",
+    platform: "twitter",
+  },
+  {
+    client: "Lisa Thompson",
+    title: "A Symphony of Sweet Flavors!",
+    comment:
+      "Goldis Cake masters the art of blending flavors to perfection. With every forkful, you're transported to a world of sweetness and delight. A must-try for any dessert enthusiast!",
+    platform: "twitter",
+  },
+  {
+    client: "Sarah Johnson",
+    title: "Deliciously Decadent Treat!",
+    comment:
+      "Goldis Cake offers a delightful range of flavors and textures. Each bite is like a symphony of sweetness that melts in your mouth. Simply divine!",
+    platform: "Facebook",
+  },
+  {
+    client: "Michael Chang",
+    title: "Sensational Sweetness in Every Slice!",
+    comment:
+      "Goldis Cake truly knows how to satisfy a sweet tooth. From the moist sponge to the rich frosting, every slice is bursting with flavor. It's a dessert lover's dream come true!",
+    platform: "twitter",
+  },
+  {
+    client: "Emily Rodriguez",
+    title: "Heavenly Indulgence on a Plate!",
+    comment:
+      "Indulging in Goldis Cake is like taking a trip to dessert paradise. The cakes are moist, the toppings are scrumptious, and every bite is a taste sensation. Pure bliss!",
+    platform: "Facebook",
+  },
+  {
+    client: "David Patel",
+    title: "Irresistibly Tempting Treats!",
+    comment:
+      "Goldis Cake crafts cakes that are impossible to resist. From the first glance to the last crumb, their creations are a feast for the eyes and the taste buds. Simply irresistible!",
+    platform: "Facebook",
+  },
+  {
+    client: "Lisa Thompson",
+    title: "A Symphony of Sweet Flavors!",
+    comment:
+      "Goldis Cake masters the art of blending flavors to perfection. With every forkful, you're transported to a world of sweetness and delight. A must-try for any dessert enthusiast!",
+    platform: "twitter",
+  },
+  {
+    client: "Kevin Nguyen",
+    title: "Pure Perfection in Every Bite!",
+    comment:
+      "Goldis Cake elevates the cake experience to new heights. Each creation is a masterpiece of taste and texture, leaving you craving more. It's perfection on a plate!",
+    platform: "twitter",
+  },
+];
